+++ src/terminal.go
@@ -1,8 +1,10 @@
 package fzf
 
 import (
+	"bufio"
 	"bytes"
 	"fmt"
+	"io"
 	"os"
 	"os/signal"
 	"regexp"
@@ -15,8 +17,6 @@ import (
 
 	"github.com/junegunn/fzf/src/tui"
 	"github.com/junegunn/fzf/src/util"
-
-	"github.com/junegunn/go-runewidth"
 )
 
 // import "github.com/pkg/profile"
@@ -42,6 +42,16 @@ type previewer struct {
 	enabled bool
 }
 
+type itemLine struct {
+	current  bool
+	selected bool
+	label    string
+	width    int
+	result   Result
+}
+
+var emptyLine = itemLine{}
+
 // Terminal represents terminal input/output
 type Terminal struct {
 	initDelay  time.Duration
@@ -50,6 +60,8 @@ type Terminal struct {
 	reverse    bool
 	hscroll    bool
 	hscrollOff int
+	wordRubout string
+	wordNext   string
 	cx         int
 	cy         int
 	offset     int
@@ -69,11 +81,12 @@ type Terminal struct {
 	header     []string
 	header0    []string
 	ansi       bool
+	tabstop    int
 	margin     [4]sizeSpec
 	strong     tui.Attr
-	window     *tui.Window
-	bwindow    *tui.Window
-	pwindow    *tui.Window
+	window     tui.Window
+	bwindow    tui.Window
+	pwindow    tui.Window
 	count      int
 	progress   int
 	reading    bool
@@ -89,10 +102,12 @@ type Terminal struct {
 	eventBox   *util.EventBox
 	mutex      sync.Mutex
 	initFunc   func()
+	prevLines  []itemLine
 	suppress   bool
 	startChan  chan bool
 	slab       *util.Slab
 	theme      *tui.ColorTheme
+	tui        tui.Renderer
 }
 
 type selectedItem struct {
@@ -115,7 +130,6 @@ func (a byTimeOrder) Less(i, j int) bool {
 }
 
 var _spinner = []string{`-`, `\`, `|`, `/`, `-`, `\`, `|`, `/`}
-var _runeWidths = make(map[rune]int)
 var _tabStop int
 
 const (
@@ -172,6 +186,8 @@ const (
 	actUp
 	actPageUp
 	actPageDown
+	actHalfPageUp
+	actHalfPageDown
 	actJump
 	actJumpAccept
 	actPrintQuery
@@ -190,6 +206,7 @@ const (
 func defaultKeymap() map[int]actionType {
 	keymap := make(map[int]actionType)
 	keymap[tui.Invalid] = actInvalid
+	keymap[tui.Resize] = actClearScreen
 	keymap[tui.CtrlA] = actBeginningOfLine
 	keymap[tui.CtrlB] = actBackwardChar
 	keymap[tui.CtrlC] = actAbort
@@ -246,7 +263,6 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 	} else {
 		header = reverseStringArray(opts.Header)
 	}
-	_tabStop = opts.Tabstop
 	var delay time.Duration
 	if opts.Tac {
 		delay = initialDelayTac
@@ -261,6 +277,32 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 	if !opts.Bold {
 		strongAttr = tui.AttrRegular
 	}
+	var renderer tui.Renderer
+	if opts.Height.size > 0 {
+		maxHeightFunc := func(termHeight int) int {
+			var maxHeight int
+			if opts.Height.percent {
+				maxHeight = util.Min(termHeight,
+					util.Max(int(opts.Height.size*float64(termHeight)/100.0), opts.MinHeight))
+			} else {
+				maxHeight = util.Min(termHeight, int(opts.Height.size))
+			}
+			if opts.InlineInfo {
+				return util.Max(maxHeight, minHeight-1)
+			}
+			return util.Max(maxHeight, minHeight)
+		}
+		renderer = tui.NewLightRenderer(opts.Theme, opts.Black, opts.Mouse, opts.Tabstop, maxHeightFunc)
+	} else {
+		renderer = tui.NewFullscreenRenderer(opts.Theme, opts.Black, opts.Mouse)
+	}
+	wordRubout := "[^[:alnum:]][[:alnum:]]"
+	wordNext := "[[:alnum:]][^[:alnum:]]|(.$)"
+	if opts.FileWord {
+		sep := regexp.QuoteMeta(string(os.PathSeparator))
+		wordRubout = fmt.Sprintf("%s[^%s]", sep, sep)
+		wordNext = fmt.Sprintf("[^%s]%s|(.$)", sep, sep)
+	}
 	return &Terminal{
 		initDelay:  delay,
 		inlineInfo: opts.InlineInfo,
@@ -268,6 +310,8 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 		reverse:    opts.Reverse,
 		hscroll:    opts.Hscroll,
 		hscrollOff: opts.HscrollOff,
+		wordRubout: wordRubout,
+		wordNext:   wordNext,
 		cx:         len(input),
 		cy:         0,
 		offset:     0,
@@ -289,6 +333,7 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 		header:     header,
 		header0:    header,
 		ansi:       opts.Ansi,
+		tabstop:    opts.Tabstop,
 		reading:    true,
 		jumping:    jumpDisabled,
 		jumpLabels: opts.JumpLabels,
@@ -305,9 +350,8 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 		slab:       util.MakeSlab(slab16Size, slab32Size),
 		theme:      opts.Theme,
 		startChan:  make(chan bool, 1),
-		initFunc: func() {
-			tui.Init(opts.Theme, opts.Black, opts.Mouse)
-		}}
+		tui:        renderer,
+		initFunc:   func() { renderer.Init() }}
 }
 
 // Input returns current query string
@@ -400,22 +444,10 @@ func (t *Terminal) sortSelected() []selectedItem {
 	return sels
 }
 
-func runeWidth(r rune, prefixWidth int) int {
-	if r == '\t' {
-		return _tabStop - prefixWidth%_tabStop
-	} else if w, found := _runeWidths[r]; found {
-		return w
-	} else {
-		w := runewidth.RuneWidth(r)
-		_runeWidths[r] = w
-		return w
-	}
-}
-
-func displayWidth(runes []rune) int {
+func (t *Terminal) displayWidth(runes []rune) int {
 	l := 0
 	for _, r := range runes {
-		l += runeWidth(r, l)
+		l += util.RuneWidth(r, l, t.tabstop)
 	}
 	return l
 }
@@ -436,9 +468,10 @@ func calculateSize(base int, size sizeSpec, margin int, minSize int) int {
 }
 
 func (t *Terminal) resizeWindows() {
-	screenWidth := tui.MaxX()
-	screenHeight := tui.MaxY()
+	screenWidth := t.tui.MaxX()
+	screenHeight := t.tui.MaxY()
 	marginInt := [4]int{}
+	t.prevLines = make([]itemLine, screenHeight)
 	for idx, sizeSpec := range t.margin {
 		if sizeSpec.percent {
 			var max float64
@@ -486,43 +519,55 @@ func (t *Terminal) resizeWindows() {
 	height := screenHeight - marginInt[0] - marginInt[2]
 	if t.isPreviewEnabled() {
 		createPreviewWindow := func(y int, x int, w int, h int) {
-			t.bwindow = tui.NewWindow(y, x, w, h, true)
-			t.pwindow = tui.NewWindow(y+1, x+2, w-4, h-2, false)
+			t.bwindow = t.tui.NewWindow(y, x, w, h, true)
+			pwidth := w - 4
+			// ncurses auto-wraps the line when the cursor reaches the right-end of
+			// the window. To prevent unintended line-wraps, we use the width one
+			// column larger than the desired value.
+			if !t.preview.wrap && t.tui.DoesAutoWrap() {
+				pwidth += 1
+			}
+			t.pwindow = t.tui.NewWindow(y+1, x+2, pwidth, h-2, false)
 		}
 		switch t.preview.position {
 		case posUp:
 			pheight := calculateSize(height, t.preview.size, minHeight, 3)
-			t.window = tui.NewWindow(
+			t.window = t.tui.NewWindow(
 				marginInt[0]+pheight, marginInt[3], width, height-pheight, false)
 			createPreviewWindow(marginInt[0], marginInt[3], width, pheight)
 		case posDown:
 			pheight := calculateSize(height, t.preview.size, minHeight, 3)
-			t.window = tui.NewWindow(
+			t.window = t.tui.NewWindow(
 				marginInt[0], marginInt[3], width, height-pheight, false)
 			createPreviewWindow(marginInt[0]+height-pheight, marginInt[3], width, pheight)
 		case posLeft:
 			pwidth := calculateSize(width, t.preview.size, minWidth, 5)
-			t.window = tui.NewWindow(
+			t.window = t.tui.NewWindow(
 				marginInt[0], marginInt[3]+pwidth, width-pwidth, height, false)
 			createPreviewWindow(marginInt[0], marginInt[3], pwidth, height)
 		case posRight:
 			pwidth := calculateSize(width, t.preview.size, minWidth, 5)
-			t.window = tui.NewWindow(
+			t.window = t.tui.NewWindow(
 				marginInt[0], marginInt[3], width-pwidth, height, false)
 			createPreviewWindow(marginInt[0], marginInt[3]+width-pwidth, pwidth, height)
 		}
 	} else {
-		t.window = tui.NewWindow(
+		t.window = t.tui.NewWindow(
 			marginInt[0],
 			marginInt[3],
 			width,
 			height, false)
 	}
+	if !t.tui.IsOptimized() && t.theme != nil && t.theme.HasBg() {
+		for i := 0; i < t.window.Height(); i++ {
+			t.window.MoveAndClear(i, 0)
+		}
+	}
 }
 
 func (t *Terminal) move(y int, x int, clear bool) {
 	if !t.reverse {
-		y = t.window.Height - y - 1
+		y = t.window.Height() - y - 1
 	}
 
 	if clear {
@@ -533,7 +578,7 @@ func (t *Terminal) move(y int, x int, clear bool) {
 }
 
 func (t *Terminal) placeCursor() {
-	t.move(0, displayWidth([]rune(t.prompt))+displayWidth(t.input[:t.cx]), false)
+	t.move(0, t.displayWidth([]rune(t.prompt))+t.displayWidth(t.input[:t.cx]), false)
 }
 
 func (t *Terminal) printPrompt() {
@@ -544,7 +589,7 @@ func (t *Terminal) printPrompt() {
 
 func (t *Terminal) printInfo() {
 	if t.inlineInfo {
-		t.move(0, displayWidth([]rune(t.prompt))+displayWidth(t.input)+1, true)
+		t.move(0, t.displayWidth([]rune(t.prompt))+t.displayWidth(t.input)+1, true)
 		if t.reading {
 			t.window.CPrint(tui.ColSpinner, t.strong, " < ")
 		} else {
@@ -581,7 +626,7 @@ func (t *Terminal) printHeader() {
 	if len(t.header) == 0 {
 		return
 	}
-	max := t.window.Height
+	max := t.window.Height()
 	var state *ansiState
 	for idx, lineStr := range t.header {
 		line := idx + 2
@@ -599,7 +644,7 @@ func (t *Terminal) printHeader() {
 
 		t.move(line, 2, true)
 		t.printHighlighted(&Result{item: item},
-			tui.AttrRegular, tui.ColHeader, tui.ColDefault, false)
+			tui.AttrRegular, tui.ColHeader, tui.ColDefault, false, false)
 	}
 }
 
@@ -608,19 +653,25 @@ func (t *Terminal) printList() {
 
 	maxy := t.maxItems()
 	count := t.merger.Length() - t.offset
-	for i := 0; i < maxy; i++ {
+	for j := 0; j < maxy; j++ {
+		i := j
+		if !t.reverse {
+			i = maxy - 1 - j
+		}
 		line := i + 2 + len(t.header)
 		if t.inlineInfo {
 			line--
 		}
-		t.move(line, 0, true)
 		if i < count {
-			t.printItem(t.merger.Get(i+t.offset), i, i == t.cy-t.offset)
+			t.printItem(t.merger.Get(i+t.offset), line, i, i == t.cy-t.offset)
+		} else if t.prevLines[i] != emptyLine {
+			t.prevLines[i] = emptyLine
+			t.move(line, 0, true)
 		}
 	}
 }
 
-func (t *Terminal) printItem(result *Result, i int, current bool) {
+func (t *Terminal) printItem(result *Result, line int, i int, current bool) {
 	item := result.item
 	_, selected := t.selected[item.Index()]
 	label := " "
@@ -633,6 +684,19 @@ func (t *Terminal) printItem(result *Result, i int, current bool) {
 	} else if current {
 		label = ">"
 	}
+
+	// Avoid unnecessary redraw
+	newLine := itemLine{current: current, selected: selected, label: label, result: *result, width: 0}
+	prevLine := t.prevLines[i]
+	if prevLine.current == newLine.current &&
+		prevLine.selected == newLine.selected &&
+		prevLine.label == newLine.label &&
+		prevLine.result == newLine.result {
+		return
+	}
+
+	// Optimized renderer can simply erase to the end of the window
+	t.move(line, 0, t.tui.IsOptimized())
 	t.window.CPrint(tui.ColCursor, t.strong, label)
 	if current {
 		if selected {
@@ -640,33 +704,40 @@ func (t *Terminal) printItem(result *Result, i int, current bool) {
 		} else {
 			t.window.CPrint(tui.ColCurrent, t.strong, " ")
 		}
-		t.printHighlighted(result, t.strong, tui.ColCurrent, tui.ColCurrentMatch, true)
+		newLine.width = t.printHighlighted(result, t.strong, tui.ColCurrent, tui.ColCurrentMatch, true, true)
 	} else {
 		if selected {
 			t.window.CPrint(tui.ColSelected, t.strong, ">")
 		} else {
 			t.window.Print(" ")
 		}
-		t.printHighlighted(result, 0, tui.ColNormal, tui.ColMatch, false)
+		newLine.width = t.printHighlighted(result, 0, tui.ColNormal, tui.ColMatch, false, true)
 	}
+	if !t.tui.IsOptimized() {
+		fillSpaces := prevLine.width - newLine.width
+		if fillSpaces > 0 {
+			t.window.Print(strings.Repeat(" ", fillSpaces))
+		}
+	}
+	t.prevLines[i] = newLine
 }
 
-func trimRight(runes []rune, width int) ([]rune, int) {
+func (t *Terminal) trimRight(runes []rune, width int) ([]rune, int) {
 	// We start from the beginning to handle tab characters
 	l := 0
 	for idx, r := range runes {
-		l += runeWidth(r, l)
-		if idx > 0 && l > width {
+		l += util.RuneWidth(r, l, t.tabstop)
+		if l > width {
 			return runes[:idx], len(runes) - idx
 		}
 	}
 	return runes, 0
 }
 
-func displayWidthWithLimit(runes []rune, prefixWidth int, limit int) int {
+func (t *Terminal) displayWidthWithLimit(runes []rune, prefixWidth int, limit int) int {
 	l := 0
 	for _, r := range runes {
-		l += runeWidth(r, l+prefixWidth)
+		l += util.RuneWidth(r, l+prefixWidth, t.tabstop)
 		if l > limit {
 			// Early exit
 			return l
@@ -675,35 +746,28 @@ func displayWidthWithLimit(runes []rune, prefixWidth int, limit int) int {
 	return l
 }
 
-func trimLeft(runes []rune, width int) ([]rune, int32) {
+func (t *Terminal) trimLeft(runes []rune, width int) ([]rune, int32) {
 	if len(runes) > maxDisplayWidthCalc && len(runes) > width {
 		trimmed := len(runes) - width
 		return runes[trimmed:], int32(trimmed)
 	}
 
-	currentWidth := displayWidth(runes)
+	currentWidth := t.displayWidth(runes)
 	var trimmed int32
 
 	for currentWidth > width && len(runes) > 0 {
 		runes = runes[1:]
 		trimmed++
-		currentWidth = displayWidthWithLimit(runes, 2, width)
+		currentWidth = t.displayWidthWithLimit(runes, 2, width)
 	}
 	return runes, trimmed
 }
 
-func overflow(runes []rune, max int) bool {
-	l := 0
-	for _, r := range runes {
-		l += runeWidth(r, l)
-		if l > max {
-			return true
-		}
-	}
-	return false
+func (t *Terminal) overflow(runes []rune, max int) bool {
+	return t.displayWidthWithLimit(runes, 0, max) > max
 }
 
-func (t *Terminal) printHighlighted(result *Result, attr tui.Attr, col1 tui.ColorPair, col2 tui.ColorPair, current bool) {
+func (t *Terminal) printHighlighted(result *Result, attr tui.Attr, col1 tui.ColorPair, col2 tui.ColorPair, current bool, match bool) int {
 	item := result.item
 
 	// Overflow
@@ -711,7 +775,7 @@ func (t *Terminal) printHighlighted(result *Result, attr tui.Attr, col1 tui.Colo
 	copy(text, item.text.ToRunes())
 	matchOffsets := []Offset{}
 	var pos *[]int
-	if t.merger.pattern != nil {
+	if match && t.merger.pattern != nil {
 		_, matchOffsets, pos = t.merger.pattern.MatchItem(item, true, t.slab)
 	}
 	charOffsets := matchOffsets
@@ -729,22 +793,23 @@ func (t *Terminal) printHighlighted(result *Result, attr tui.Attr, col1 tui.Colo
 	}
 
 	offsets := result.colorOffsets(charOffsets, t.theme, col2, attr, current)
-	maxWidth := t.window.Width - 3
+	maxWidth := t.window.Width() - 3
 	maxe = util.Constrain(maxe+util.Min(maxWidth/2-2, t.hscrollOff), 0, len(text))
-	if overflow(text, maxWidth) {
+	displayWidth := t.displayWidthWithLimit(text, 0, maxWidth)
+	if displayWidth > maxWidth {
 		if t.hscroll {
 			// Stri..
-			if !overflow(text[:maxe], maxWidth-2) {
-				text, _ = trimRight(text, maxWidth-2)
+			if !t.overflow(text[:maxe], maxWidth-2) {
+				text, _ = t.trimRight(text, maxWidth-2)
 				text = append(text, []rune("..")...)
 			} else {
 				// Stri..
-				if overflow(text[maxe:], 2) {
+				if t.overflow(text[maxe:], 2) {
 					text = append(text[:maxe], []rune("..")...)
 				}
 				// ..ri..
 				var diff int32
-				text, diff = trimLeft(text, maxWidth-2)
+				text, diff = t.trimLeft(text, maxWidth-2)
 
 				// Transform offsets
 				for idx, offset := range offsets {
@@ -758,7 +823,7 @@ func (t *Terminal) printHighlighted(result *Result, attr tui.Attr, col1 tui.Colo
 				text = append([]rune(".."), text...)
 			}
 		} else {
-			text, _ = trimRight(text, maxWidth-2)
+			text, _ = t.trimRight(text, maxWidth-2)
 			text = append(text, []rune("..")...)
 
 			for idx, offset := range offsets {
@@ -776,11 +841,11 @@ func (t *Terminal) printHighlighted(result *Result, attr tui.Attr, col1 tui.Colo
 		b := util.Constrain32(offset.offset[0], index, maxOffset)
 		e := util.Constrain32(offset.offset[1], index, maxOffset)
 
-		substr, prefixWidth = processTabs(text[index:b], prefixWidth)
+		substr, prefixWidth = t.processTabs(text[index:b], prefixWidth)
 		t.window.CPrint(col1, attr, substr)
 
 		if b < e {
-			substr, prefixWidth = processTabs(text[b:e], prefixWidth)
+			substr, prefixWidth = t.processTabs(text[b:e], prefixWidth)
 			t.window.CPrint(offset.color, offset.attr, substr)
 		}
 
@@ -790,9 +855,10 @@ func (t *Terminal) printHighlighted(result *Result, attr tui.Attr, col1 tui.Colo
 		}
 	}
 	if index < maxOffset {
-		substr, _ = processTabs(text[index:], prefixWidth)
+		substr, _ = t.processTabs(text[index:], prefixWidth)
 		t.window.CPrint(col1, attr, substr)
 	}
+	return displayWidth
 }
 
 func numLinesMax(str string, max int) int {
@@ -813,37 +879,66 @@ func (t *Terminal) printPreview() {
 		return
 	}
 	t.pwindow.Erase()
-	skip := t.previewer.offset
-	extractColor(t.previewer.text, nil, func(str string, ansi *ansiState) bool {
-		if skip > 0 {
-			newlines := numLinesMax(str, skip)
-			if skip <= newlines {
-				for i := 0; i < skip; i++ {
-					str = str[strings.Index(str, "\n")+1:]
+
+	maxWidth := t.pwindow.Width()
+	if t.tui.DoesAutoWrap() {
+		maxWidth -= 1
+	}
+	reader := bufio.NewReader(strings.NewReader(t.previewer.text))
+	lineNo := -t.previewer.offset
+	for {
+		line, err := reader.ReadString('\n')
+		eof := err == io.EOF
+		if !eof {
+			line = line[:len(line)-1]
+		}
+		lineNo++
+		if lineNo > t.pwindow.Height() {
+			break
+		} else if lineNo > 0 {
+			var fillRet tui.FillReturn
+			extractColor(line, nil, func(str string, ansi *ansiState) bool {
+				trimmed := []rune(str)
+				if !t.preview.wrap {
+					trimmed, _ = t.trimRight(trimmed, maxWidth-t.pwindow.X())
 				}
-				skip = 0
-			} else {
-				skip -= newlines
-				return true
+				str, _ = t.processTabs(trimmed, 0)
+				if ansi != nil && ansi.colored() {
+					fillRet = t.pwindow.CFill(ansi.fg, ansi.bg, ansi.attr, str)
+				} else {
+					fillRet = t.pwindow.Fill(str)
+				}
+				return fillRet == tui.FillContinue
+			})
+			switch fillRet {
+			case tui.FillNextLine:
+				continue
+			case tui.FillSuspend:
+				break
 			}
+			t.pwindow.Fill("\n")
 		}
-		if ansi != nil && ansi.colored() {
-			return t.pwindow.CFill(str, ansi.fg, ansi.bg, ansi.attr)
+		if eof {
+			break
 		}
-		return t.pwindow.Fill(str)
-	})
-	if t.previewer.offset > 0 {
+	}
+	t.pwindow.FinishFill()
+	if t.previewer.lines > t.pwindow.Height() {
 		offset := fmt.Sprintf("%d/%d", t.previewer.offset+1, t.previewer.lines)
-		t.pwindow.Move(0, t.pwindow.Width-len(offset))
+		pos := t.pwindow.Width() - len(offset)
+		if t.tui.DoesAutoWrap() {
+			pos -= 1
+		}
+		t.pwindow.Move(0, pos)
 		t.pwindow.CPrint(tui.ColInfo, tui.Reverse, offset)
 	}
 }
 
-func processTabs(runes []rune, prefixWidth int) (string, int) {
+func (t *Terminal) processTabs(runes []rune, prefixWidth int) (string, int) {
 	var strbuf bytes.Buffer
 	l := prefixWidth
 	for _, r := range runes {
-		w := runeWidth(r, l)
+		w := util.RuneWidth(r, l, t.tabstop)
 		l += w
 		if r == '\t' {
 			strbuf.WriteString(strings.Repeat(" ", w))
@@ -866,9 +961,9 @@ func (t *Terminal) printAll() {
 func (t *Terminal) refresh() {
 	if !t.suppress {
 		if t.isPreviewEnabled() {
-			tui.RefreshWindows([]*tui.Window{t.bwindow, t.pwindow, t.window})
+			t.tui.RefreshWindows([]tui.Window{t.bwindow, t.pwindow, t.window})
 		} else {
-			tui.RefreshWindows([]*tui.Window{t.window})
+			t.tui.RefreshWindows([]tui.Window{t.window})
 		}
 	}
 }
@@ -990,9 +1085,9 @@ func (t *Terminal) executeCommand(template string, items []*Item) {
 	cmd.Stdin = os.Stdin
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
-	tui.Pause()
+	t.tui.Pause()
 	cmd.Run()
-	if tui.Resume() {
+	if t.tui.Resume() {
 		t.printAll()
 	}
 	t.refresh()
@@ -1139,11 +1234,11 @@ func (t *Terminal) Loop() {
 					case reqRefresh:
 						t.suppress = false
 					case reqRedraw:
-						tui.Clear()
-						tui.Refresh()
+						t.tui.Clear()
+						t.tui.Refresh()
 						t.printAll()
 					case reqClose:
-						tui.Close()
+						t.tui.Close()
 						if t.output() {
 							exit(exitOk)
 						}
@@ -1156,11 +1251,11 @@ func (t *Terminal) Loop() {
 					case reqPreviewRefresh:
 						t.printPreview()
 					case reqPrintQuery:
-						tui.Close()
+						t.tui.Close()
 						t.printer(string(t.input))
 						exit(exitOk)
 					case reqQuit:
-						tui.Close()
+						t.tui.Close()
 						exit(exitInterrupt)
 					}
 				}
@@ -1173,7 +1268,7 @@ func (t *Terminal) Loop() {
 
 	looping := true
 	for looping {
-		event := tui.GetChar()
+		event := t.tui.GetChar()
 
 		t.mutex.Lock()
 		previousInput := t.input
@@ -1265,11 +1360,11 @@ func (t *Terminal) Loop() {
 				}
 			case actPreviewPageUp:
 				if t.isPreviewEnabled() {
-					scrollPreview(-t.pwindow.Height)
+					scrollPreview(-t.pwindow.Height())
 				}
 			case actPreviewPageDown:
 				if t.isPreviewEnabled() {
-					scrollPreview(t.pwindow.Height)
+					scrollPreview(t.pwindow.Height())
 				}
 			case actBeginningOfLine:
 				t.cx = 0
@@ -1378,7 +1473,7 @@ func (t *Terminal) Loop() {
 				}
 			case actBackwardKillWord:
 				if t.cx > 0 {
-					t.rubout("[^[:alnum:]][[:alnum:]]")
+					t.rubout(t.wordRubout)
 				}
 			case actYank:
 				suffix := copySlice(t.input[t.cx:])
@@ -1390,6 +1485,12 @@ func (t *Terminal) Loop() {
 			case actPageDown:
 				t.vmove(-(t.maxItems() - 1))
 				req(reqList)
+			case actHalfPageUp:
+				t.vmove(t.maxItems() / 2)
+				req(reqList)
+			case actHalfPageDown:
+				t.vmove(-(t.maxItems() / 2))
+				req(reqList)
 			case actJump:
 				t.jumping = jumpEnabled
 				req(reqJump)
@@ -1397,12 +1498,12 @@ func (t *Terminal) Loop() {
 				t.jumping = jumpAcceptEnabled
 				req(reqJump)
 			case actBackwardWord:
-				t.cx = findLastMatch("[^[:alnum:]][[:alnum:]]", string(t.input[:t.cx])) + 1
+				t.cx = findLastMatch(t.wordRubout, string(t.input[:t.cx])) + 1
 			case actForwardWord:
-				t.cx += findFirstMatch("[[:alnum:]][^[:alnum:]]|(.$)", string(t.input[t.cx:])) + 1
+				t.cx += findFirstMatch(t.wordNext, string(t.input[t.cx:])) + 1
 			case actKillWord:
 				ncx := t.cx +
-					findFirstMatch("[[:alnum:]][^[:alnum:]]|(.$)", string(t.input[t.cx:])) + 1
+					findFirstMatch(t.wordNext, string(t.input[t.cx:])) + 1
 				if ncx > t.cx {
 					t.yanked = copySlice(t.input[t.cx:ncx])
 					t.input = append(t.input[:t.cx], t.input[ncx:]...)
@@ -1443,11 +1544,11 @@ func (t *Terminal) Loop() {
 						scrollPreview(-me.S)
 					}
 				} else if t.window.Enclose(my, mx) {
-					mx -= t.window.Left
-					my -= t.window.Top
-					mx = util.Constrain(mx-displayWidth([]rune(t.prompt)), 0, len(t.input))
+					mx -= t.window.Left()
+					my -= t.window.Top()
+					mx = util.Constrain(mx-t.displayWidth([]rune(t.prompt)), 0, len(t.input))
 					if !t.reverse {
-						my = t.window.Height - my - 1
+						my = t.window.Height() - my - 1
 					}
 					min := 2 + len(t.header)
 					if t.inlineInfo {
@@ -1559,7 +1660,7 @@ func (t *Terminal) vset(o int) bool {
 }
 
 func (t *Terminal) maxItems() int {
-	max := t.window.Height - 2 - len(t.header)
+	max := t.window.Height() - 2 - len(t.header)
 	if t.inlineInfo {
 		max++
 	}